Adds healthchecks
Fixes #3

The trick to making this work is the `service.alpha.kubernetes.io/tolerate-unready-endpoints`
annotation on the discovery service. This makes the service return all
the matching endpoints even before they are marked as ready.

This annotation will eventuly be replaced by `publishNotReadyAddresses`
but for now (k8s v1.9) the annotation is still needed.